pyproject.toml
[project]
name = "lasttoknow"
version = "0.2.0"
description = "🔔 Never miss what matters in tech. Track packages, releases, trends — and get briefed like a CTO."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Dinesh Karakambaka", email = "kdineshkvkl@gmail.com" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    "Environment :: Console",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.15.0",
    "rich>=13.0.0",
    "google-adk>=1.23.0",
    "litellm<=1.70.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
lasttoknow = "lasttoknow.cli:app"

[project.urls]
Homepage = "https://github.com/dineshkrishna9999/lasttoknow"
Repository = "https://github.com/dineshkrishna9999/lasttoknow"
Issues = "https://github.com/dineshkrishna9999/lasttoknow/issues"
Documentation = "https://github.com/dineshkrishna9999/lasttoknow#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lasttoknow"]

# ──────────────────────────────────────────────
# Poe the Poet — Task Runner
# ──────────────────────────────────────────────
[tool.poe]
include = "poe.toml"

# ──────────────────────────────────────────────
# Ruff — Linter & Formatter
# ──────────────────────────────────────────────
[tool.ruff]
line-length = 120
target-version = "py311"
fix = true
src = ["src", "tests"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "W",     # pycodestyle warnings
    "B",     # flake8-bugbear
    "Q",     # flake8-quotes
    "I",     # isort
    "ASYNC", # flake8-async
    "T20",   # flake8-print
    "UP",    # pyupgrade
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
    "S",     # flake8-bandit (security)
    "PTH",   # flake8-use-pathlib
    "TCH",   # flake8-type-checking
    "N",     # pep8-naming
]
ignore = [
    "E501",   # line length (handled by formatter)
    "S101",   # assert in tests is fine
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports OK in __init__
"tests/**" = ["S101"]     # assert is fine in tests

[tool.ruff.lint.isort]
known-first-party = ["lasttoknow"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query", "fastapi.Path", "fastapi.Body"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"unittest".msg = "Use pytest instead of unittest."

# ──────────────────────────────────────────────
# Mypy — Static Type Checker
# ──────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true

# Third-party libs that don't ship type stubs
[[tool.mypy.overrides]]
module = [
    "google.adk.*",
    "google.genai.*",
    "litellm.*",
]
ignore_missing_imports = true

# ──────────────────────────────────────────────
# Pyright — Static Type Checker
# ──────────────────────────────────────────────
[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.11"
typeCheckingMode = "standard"
reportUnnecessaryIsInstance = false
reportMissingTypeStubs = false

# ──────────────────────────────────────────────
# Pytest — Testing
# ──────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",          # show summary of all non-passing tests
    "-q",           # quieter output
    "--strict-markers",
    "--tb=short",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
]

# ──────────────────────────────────────────────
# Coverage
# ──────────────────────────────────────────────
[tool.coverage.run]
source = ["src/lasttoknow"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "@overload",
]

# ──────────────────────────────────────────────
# Dev Dependencies
# ──────────────────────────────────────────────
[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.10.0",
    "pyright>=1.1.378",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=3.6.1",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "poethepoet>=0.30.0",
    "pre-commit>=4.0.0",
]
